requirements.txt
absl-py==0.7.0
alabaster==0.7.12
albumentations==0.1.11
allennlp==0.9.0
apex==0.1
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==1.5.0
blis==0.2.4
boto==2.49.0
boto3==1.9.83
botocore==1.12.83
bz2file==0.98
cachetools==4.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
codecov==2.0.15
conda==4.5.12
conda-build==3.16.2
conllu==1.3.1
coverage==4.5.1
cryptography==2.3.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dask==1.1.0
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
docutils==0.14
editdistance==0.5.3
entrypoints==0.2.3
filelock==3.0.10
flake8==3.6.0
flaky==3.6.1
Flask==1.0.2
Flask-Cors==3.0.8
ftfy==5.6
future==0.17.1
gast==0.2.2
gensim==3.7.0
geojson==2.4.1
gevent==1.4.0
glob2==0.6
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.16.1
h5py==2.9.0
html5lib==0.9999999
hypothesis==3.82.1
idna==2.7
imageio==2.4.1
imagesize==1.1.0
imgaug==0.2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonnet==0.14.0
jsonpickle==1.2
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kaggle==1.5.1.1
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
libarchive-c==2.8
lmdb==0.94
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.0
nibabel==2.3.3
ninja==1.8.2.post2
nltk==3.3
notebook==5.7.0
numpy==1.18.0
numpydoc==0.9.2
nvidia-dali==0.4.1
oauthlib==3.1.0
onnx==1.3.0
opencv-python==4.0.0.21
opencv-python-headless==4.0.0.21
opt-einsum==3.1.0
overrides==2.7.0
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.3.1
patsy==0.5.1
pbr==5.1.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
Pillow-SIMD==5.3.0.post0
pkginfo==1.4.2
plac==0.9.6
pluggy==0.8.0
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pydicom==1.2.2
pydot==1.2.3
pyflakes==2.0.0
Pygments==2.2.0
pygpu==0.7.6
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==3.9.3
pytest-cov==2.6.0
pytest-pythonpath==0.7.3
python-dateutil==2.7.5
python-nvd3==0.15.0
python-slugify==1.2.6
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
regex==2018.1.10
requests==2.22.0
requests-oauthlib==1.3.0
responses==0.10.9
revtok==0.0.3
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.1.13
sacrebleu==1.2.10
sacremoses==0.0.5
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.9.0
segmentation-models==0.1.2
Send2Trash==1.5.0
sentencepiece==0.1.85
Shapely==1.6.4.post2
six==1.11.0
smart-open==1.8.0
snowballstemmer==1.2.1
spacy==2.1.9
Sphinx==1.8.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
srsly==1.0.0
statsmodels==0.9.0
tabulate==0.8.2
tensorboard==2.0.2
tensorboardX==1.6
tensorflow==1.7.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==7.0.8
tifffile==2019.1.4
toml==0.10.0
toolz==0.9.0
torch==1.3.1
torchtext==0.3.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.6.6
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
wasabi==0.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
word2number==1.1
wrapt==1.11.2
